requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
pillow==10.1.0
opencv-python==4.8.1.78
aiomysql==0.2.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1
python-decouple==3.8
pydantic==2.5.0
python-magic==0.4.27
python-magic-bin==0.4.14
jinja2==3.1.2
aiofiles==23.2.0
